Adding plugin logic to DatadockThread.
git-svn-id: svn://oss.beweb.dbc.dk/sync_search/opensearch/trunk@3379 51cac136-f83d-432d-bb6f-125e83a60a2d

+++ src/dk/dbc/opensearch/common/pluginframework/tests/PluginResolverTest.java
@@ -62,7 +62,7 @@ public class PluginResolverTest
             System.out.print(" hep finder \n");
         }
         */
-        public String getPluginClassName( long key ) throws PluginResolverException, FileNotFoundException
+        public String getPluginClassName( int key ) throws PluginResolverException, FileNotFoundException
         {
             if (key == ( "testSubmitter"+"testFormat"+"throwException" ).hashCode())
             {
